+++ extensions/aligning_guidelines/util/collect-point.ts
